Fix bugs in g_cond_wait usage, and omx_base_sink memory leak.
The GLib user manual states that spurious wakeups are a thing that can
happen when using condition variables, and as such the caller must not
rely on the fact that a g_cond_wait returned to assume the condition
is true.

As a second bug, the OMX base sink element was missing a call to
g_omx_core_deinit() during cleanup that would lead to a significant
memory leak in the VPSS M3 core.